requirements.txt
pytest==7.4.0
selenium==4.15.0
webdriver-manager==4.0.1
allure-pytest==2.13.0
requests==2.31.0
requests==2.31.0
allure-pytest==2.13.2
python-dotenv==1.0.0
pytest==7.4.3